op-chain-ops: move top level stuff to ether package
Moves the previously top level files to a new
package called `ether` that is responsible for
doing the legacy eth migration.

The OVM_ETH ERC20 token representation of ETH
will be migrated to actual ETH that is in the
storage trie.

This makes the system eth equiavalent.
