+++ op-chain-ops/ether/addresses.go
@@ -1,4 +1,4 @@
-package state_surgery
+package ether
 
 import (
 	"bufio"
