+++ op-chain-ops/ether/cli.go
@@ -1,4 +1,4 @@
-package state_surgery
+package ether
 
 import (
 	"bytes"
