+++ op-chain-ops/ether/genesis.go
@@ -1,4 +1,4 @@
-package state_surgery
+package ether
 
 import (
 	"encoding/json"
