+++ op-chain-ops/ether/util.go
@@ -1,4 +1,4 @@
-package state_surgery
+package ether
 
 import (
 	"fmt"
